// components/ComponentMigrator/0.1.0.0/src/ts/layer3/ComponentMigratorModel.interface.ts
/**
 * ComponentMigratorModel - ComponentMigrator Component Model Interface
 * Web4 pattern: Component model following auto-discovery patterns
 */

import { Model } from './Model.interface.js';

export interface ComponentMigratorModel extends Model {
  uuid: string;
  name: string;
  origin: string;
  definition: string;
  createdAt: string;
  updatedAt: string;
}